Refactor layout of detailed forecast
Refactor the layout of the detailed forecast elements so there is even
spacing between the elements. This required pre-calculating the width
of the scrolling texts so the padding between elements can be
pre-calculated before drawing. Also refactor other related code to
create bitmaps immediately instead of when drawing, this simplifies
the drawing code significantly and ensures the bitmap creation is done
only once.