Rework agreement charges.
Apparently it is not possible to capture charges that are older than
180 days old. This required a rework of how we handle agreement charges.

The previous implementation automatically created a new charge when
the old one was captured, but this is not sufficient if the next charge
is more than 180 days into the future. The new implementation creates
charges in a separate job only when the agreement is due within 30 days.

Rework the tests and the code to work with the new flow.