Add login with email
Adds functionality for logging in using only an email. Integration with
resend for email sending functionality. Also fix subscription flow to
allow creating a subscription without having an account previously. This
required removing the current 1 kr for 1 week campaign since we have no
way of knowing if a user is new or not. If a user starts a subscription
without having an account, we ask Vipps for userinfo and create one
when returning from Vipps. Fix frontend styling to allow login with
email.