Add support for location based forecast.
This commit adds support for location based forecast. The watch will use
the new location based endpoint from the backend to get the forecast for
the closest "Type A"-region. The location-based warning is completely
independent from the other warnings, and will end up in the glance and
above all the user-selected areas.