Automatically add watch.
Change the behavior of the setup flow to allow automatically adding the
watch when the user clicks the notification link. Do this by adding a
watchKey query parameter to the link, and make sure to forward this
all the way to the account page. If the user ends up on the account page
with a watchKey-parameter, attempt to automatically add the watch.

Required reworking all the login flows as well as the vipps agreement
flow. Add watchKey to the Agreement-entity as well to be able to pick
it up when the user comes back from Vipps.